Fix size/sign conversion warnings in isabbreviation()
clang said:
xgamma.c:85:26: warning: implicit conversion changes signedness: 'int' to
      'size_t' (aka 'unsigned long') [-Wsign-conversion]
    if (strncmp (arg, s, slen) == 0) return (True);
        ~~~~~~~          ^~~~
xgamma.c:78:14: warning: implicit conversion loses integer precision: 'size_t'
      (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
    arglen = strlen (arg);
           ~ ^~~~~~~~~~~~
xgamma.c:79:12: warning: implicit conversion loses integer precision: 'size_t'
      (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
    slen = strlen (s);
         ~ ^~~~~~~~~~

+++ xgamma.c
@@ -67,10 +67,10 @@ Syntax(void)
  */
 
 static Bool 
-isabbreviation(char *arg, char *s, int minslen)
+isabbreviation(char *arg, char *s, size_t minslen)
 {
-    int arglen;
-    int slen;
+    size_t arglen;
+    size_t slen;
 
     /* exact match */
     if (strcmp (arg, s) == 0) return (True);
